Update bitlocker-recovery-guide-plan.md
The original code of GetBitLockerKeyPackageADDS.vbs script also contained the code of the GetBitLockerKeyPackage.vbs script. So, I split the code and formatted the introduction text before the code of the second script.

In the code of the first script I changed the intended name of the script: GetBitLockerKeyPackageADDS

In both scripts I put the following phrase in the usage string: "Path To Save Key Package". I'd say it is better than "Path To Saved Key Package" because before you run the script the key package has not been saved yet (the script saves the key package).