Update implement-microsoft-passport-in-your-organization.md
I am managing a server with Windows Server 2012 R2 Essentials (build 9600) and this text misled me to believe that **Turn on PIN sign-in** group policy did not affect the convenience PIN for computers running version 1607. This was upsetting because group policies relevant to the Windows Hello for Business component are not available until Windows Server 2016. After further research and experimentation, I found that **Turn on PIN sign-on** group policy object set by the server does override the **Turn on convenience PIN sign-in** local policy.