Update use-system-center-configuration-manager-to-manage-devices-with-semm.md
Running PowerShell via Configuration Manager is supported, what the script does is largely irrelevant. The statement is misleading as use of the Surface Configurator Tool in MSI format is dependant on the type of device (i.e. Surface Pro 4 or Surface book), the scripts just execute the same code in a slightly more enterprise friendly way.