Fixed broken links and typos
Fixed broken links and then typo in another link (two forward slashes although it does not affect functionality)

FYI I did remove some .md from a couple links ---- but note .md should be ok as it helps internally and externally keep links in tact.  However for some reason the process that reviews this page does not automatically filter our the .md when the page is displayed externally which causes the 404 error for the page.