Merged PR 2333: fix URL links that can't be relative due to different content sets
fix URL links that can't be relative due to different content sets
https://review.docs.microsoft.com/en-us/windows/deployment/change-history-for-deploy-windows-10?branch=fixlink is the staged version that is working. See the section "January 2017" starting with the 4th link.

https://docs.microsoft.com/en-us/windows/deployment/change-history-for-deploy-windows-10 is the live version that is broken.

The issue was that an absolute link was used but .md was included on the end of the link. At first I attempted to fix the link by making it relative, but this doesn't work because it-client contains different content sets that aren't linkable via relative links.