Add note on joining to Azure AD when device already setup
Adding an important note to the section with instructions on joining a device to Azure AD when it has already been setup. Without that disclaimer, a user going through a rather common scenario will experience a silent error where clicking the button does NOTHING leaving the user at a loss as to why things aren't working.