Consolidation on the "Summary" section
(1) The "Disk validation" section is moved up, right below the "Summary" section and is called "Prerequisites". Reason: It is so important that must not be buried. (2) Consolidated all elements of what MBR2GPT can do under a pre-existing bullet list and removed repetitions. (3) Created a list of what MBR2GPT cannot do.