fix the tag and redesign the table for "Log file"
<TD rowspan=5>setupact.log
"<TD>" tag is one of HTML tag and may be used for desigining the table. It is curious for "<TD>" tag is printed directly in the document.
And, table has four rows.
- Log file
- Phase: Location
- Description
- When to use.
In "When to use" rows, you can see some columns are none. I think the table is broken because of "<TD>" tag is sticked out