Reference Enterprise SKU at the top
Saves a lot of time with following instructions if you specify this Enterprise SKU reference.  People find out eventually that they can't run it but not before they've gotten all the way down to the steps.  Especially when you Bing search for this, and you skip the FIRST doc that specifies systems requirements.