Merged PR 6783: Added warning about custom shell leading to undeployable image
Added warning about custom shell leading to undeployable image - per request from partners, dev, and PMs.

@<Jeanie Decker> - can you take a look? I can forward you the thread that led to this request, if needed. Here's part of the thread, from Michael Niehaus:
The “hide shell” setup is really just a “RunOnce” entry that never really finishes – it might reboot the machine and run again, and eventually runs out of commands to process so at that point it exits.  So that is still using Explorer.exe as John said.

There was a question on EShell.exe too:  I believe that’s a creation of the Embedded team (I believe Suma has some background) that was later integrated into standard Windows SKUs.

Back to John’s question though:

is it fair to make the statement that setting a custom shell prior to OOBE won’t result in a deployable image?  That’s been true for a couple of releases now, and I don’t think that necessarily directly impacts the scenarios that Michael has highlighted as uses of custom shell, but I want to confirm.

Thanks,
-Michael