Edited in "sampling" into the article
In-app about page indicates that Microsoft uses "sampling", but this word does not exist on the docs page. Users have complained that the docs page does not mention anything about "sampling"; in fact, we do mention it generally but without ever using the word "sampling", so users have had a difficult time finding the section.