Fixing the default registry value set for DisableTelemetryOptInChangeNotification
This change, fixes the "Registry" section for "Diagnostic opt-in change notifications". The "GP/MDM" sections contradict the "Registry" section of the document. The default value (i.e. Enable mentioned in GP/MDM section) is 0 and not 1, setting the registry key to 1 will actually disable the notification, while the default behavior is to have it not set OR enabled (i.e. 0).