Client Management/MDM: repair broken links in oma-dm-protocol-support.md
Links to the OMA website containing OMA Device Management documents
have been broken for some time now, due to OMA website rearrangements.

This is an attempt to restore the links to the OMA website as much as possible,
given the existing references, to the documents or their directory containers.

Thanks to vlad-MSFT for providing the initial link to the website folder structure.

Closes #2759
Closes #2362