Deployment/Update: typos and whitespace corrections in waas-restart.md
During a search for the source for the Japanese version of the public page,
I noticed a typo in the sentence "Automatic Updates will notify the user that
the computer will automatically restarts in 5 minutes to complete the installation".

The word "restarts" is incorrectly written as if it was present tense 3rd person singular,
when in fact it is a Future Conditional (will do - when / if).

I also found another typo: "instllation" instead of 'installation' - now corrected.

The rest of the changes consists purely of removal of redundant blank space.