Updated Roles and Permissions
I am making these suggestion to better reflect Roles and Permissions options that are currently available in the Microsoft Store for Business tenant, as well as the text and titles that appear in the MSfB/MSfE service currently. I am proposing to remove the "Global Administrator and Billing Administration section since the "Billing Administrator" is not assignable in the current release, and I believe could lead to confusion.  I also made some other minor wording changes that I believe might make the topic more easily consumed. I hope this fork will be considered