Hardware and software requirements modification
In Hardware and software requirements Windows hypervisor was listed which feels like Hyper-V feature is the requirement. So the user has suggested that it should be mentioned that Hyper-V is not a requirement. @DaniHalfin has already communicated the information in details so I have added this info as well to make the statement clear.