Update deploy-enterprise-licenses.md
We are getting a lot of calls generated in Support because the customers use slmgr /dli or slmgr /dlv to pull the license information when they use E3/E5 activation and get the default Windows 10 Pro key which for them is a problem and think that the client has been downgraded from Ent to Pro. We need this information public until slmgr /dli or /dlv will be able to pull the E3/E5 information so we avoid more calls being generated.