Deployment/Update: netsh winsock reset command order correction
The correct order of the command `netsh winsock reset` is required.
The argument 'winsock' is the **context** and can not be rearranged.
In this context, 'reset' is the parameter to the context, not the context itself.
The correct command / context order is unchanged since Windows 2000.

Current order in the example is incorrect and needs to be amended:

-  `netsh reset winsock` does not work, it returns an error message.
-  `netsh winsock reset` is the correct order of context and parameter.

Closes #3157
Closes #3193 (if not closed already)