Update configure-windows-diagnostic-data-in-your-organization.md
Changed wording of setting diagnostic data levels via registry editor so that it is clear that if an MDM/GP policy exists, it writes over the manually set value. This avoids the erroneous confusion that the MDM/GP 'overrides' local device registry level (which is not entirely true, as the more restrictive value is taken).