USMT ScanState Syntax: hidden unescaped characters
Asterisks, backslashes or combinations of asterisk and backslash
need to be escaped for the character to migrate properly to the
docs.microsoft.com site as visible text in HTML. Github shows the
characters well enough, but the migration process does not seem to
keep the special characters through the MarkDown-to-HTML conversion.

In this PR, I have made a "best effort" attempt to resolve the missing
or malformed command examples in the "USMT ScanState Syntax" page.

Closes #2388