Windows/Privacy: URL correction update
This is a follow-up to my previous PR #3305
(Windows/Privacy: change formatting code to text).

I found one URL that I missed in my previous solution.

This is a 1 line change, making the starting and ending asterisk
show up on the page, instead of formatting the text as italics.

Addendum: For future web page documentation, it may be a good idea
to use the HTML codes `&#42;` + `&#92;` for `*` and `\` respectively,
to avoid situations where pages look OK on Github, but not translating
well to the docs.microsoft.com pages.

Ref. issue ticket #3304 (Domain misspelling)