Adding Acrolinx configuration file (#68)
Specifying that Acrolinx only checks the windows subfolder/docset. All other folders should be excluded. check only occurs after all other configuration is final (it's not yet) and when pull requests are made against the specified branch (master in this case). We are in testing only and this will not stop publishing in any way. Please contact krowley, jtidwell, or OfficeDocs-admins for information.