updated the link for msftncsi
I have updated the URL to NCSI test. From my understading, the networking blog should not be connected here with the network connection status icon. What I feel from a tech point of view, it was just a statement but mistakenly connected with the blog. It should not be connected to the network site blog. 

Additionally, we need to check with someone to bring the archived blog back online or disable the redirect. 

Problem: https://github.com/MicrosoftDocs/windows-itpro-docs/issues/4003