Microsoft Edge/kiosk mode: broken relative links
Description:

The 4 images in the subsection "Supported configuration types" are
supposed to be linked to their image files, but the relative links have
been broken by not containing the needed parent directory dots.

Proposed change:

Add the required parent directory dots to enable the links to point to
the files as intended (even if not strictly needed to begin with).

Closes #4275