resolves #4484
#4484 Page does not conform to standards

> This page, unlike other audit subcategory pages, has everything after the table of recommendations in giant font (including Feedback at the end), and does not contain any subpages for the individual events
> It also has an "In this article" panel on the right, which lists all the events.

Commit 8756c9d resulted in the page being reformatted in an odd way.

Using other audit pages as a guide, I've removed the headings that caused the events to be rendered in a giant font, and reformatted the events into a simple bulleted list. I've also added back the information on event volume and default value, to make the page conform better. I can't do much for links right now, as these pages do not exist yet.