HTML to MarkDown in hello-hybrid-aadj-sso-cert.md
This is a combined effort to alleviate a translation bug as well as
improving the MarkDown codestyle in this document, both for the English
(en-us) version of the document as well as the translated versions.

This change should in theory close the issue tickets #3451 and #3453
after the scripted translation process has been re-run on this document.

This solution is based on a user discussion in issue ticket #4589 .