Update hello-planning-guide.md + MD improvements
Description:
As agreed upon in issue ticket #4875, the line
"> This guide only applies to Windows 10, version 1511 or higher."
should be removed because it imposes an outdated build.

Thanks to mapalko for requesting this removal, and to poortom1004
for pointing out the version redundancy in the first place.

Proposed changes:
- Remove the aforementioned line, as agreed upon
- Add recommended single MarkDown quote spaces
- Replace a tab with standardized spacing
- Remove redundant white space in 3 blank lines

issue ticket closure or reference:

Closes #4875