HoloLens: move Deploy link to Quick reference
Description:
As suggested in issue ticket #5065 (Cleanly separate user from IT docs),
the link for "Deploy HoloLens in a commercial environment" should be
moved from the Guides section to the Quick reference section.

Thanks to Microsoft internal Graham Bury (grbury) for the suggestion.

Proposed change:
- move the Deploy link to the Quick reference section

issue ticket closure or reference:

Closes #5065