Use a Boot folder for case sensitive TFTP servers
I just noticed that the bootloader looks after a BCD file as follow : '\Boot\BCD'. Thus, if you are using a case sensitive TFTP server, you need to place the BCD file in a Boot folder instead of a 'boot' folder.
If you do this, you need to change all the paths specified in the BCD file, still because of the case sensitive server (otherwise, it won't find the sdi/wim files).