WHfB/Key Trust models: PIN reset availability
Description:
Based on user feedback and author verification, the PIN reset feature in
Windows Hello for Business is available in Windows 10 Professional using
different key trust models in different deployments depending on
the version of Windows 10 (1511, 1703, 1709 and 1903).

Thanks to @greytone for noting this issue in ticket #4662 .

Changes proposed:

- add important Note outlining which Windows 10 version makes the PIN
  reset feature available and the key trust model to use in each case.
  The information summary is provided by @jvsam and @mapalko, based on
  the discussion in the issue ticket page and original user feedback.

- add spacing for another important note with MD quote indentation
- add spacing for a table quote indentation
- remove HTML tag incorrectly showing up in the MarkDown preview

Additional notes:
PR content and placement of the important note is subject to change,
based on feedback from the author and MS Docs team members.

issue ticket closure or reference:

Closes #4662