Reboot CSP/ISO 8601: reboot reset date+time format
Description:

There have been repeated cases of misunderstanding the time format
needed to reset the reboot date & time schedule, lately in ticket #4810.

This is an attempt to shift the focus over to the fact that the Reboot
CSP strictly follows the ISO 8601 standard, also when resetting the
date and time for a reboot schedule.

Changes proposed:

- Add an extra sentence to inform the reader that the empty (null) value
  to reset an existing reboot schedule needs to contain zeros
  (0000-00-00T00:00:00).

- Move the sentence "Setting a null (empty) date will delete [...]" to
  a line below the example to configure, as well as the addendum,
  to focus on this fact that a format representation is required.

- Add the missing spacing in "ISO 8601" to keep its ISO name standard.

- Remove redundant white space (10 lines) at the end of the document.

issue ticket closure or reference:

Closes #4810