BitLocker and AD DS: back up recovery information
Description:
Based on user feedback in issue ticket #5159
(Specifying the parameter '-ID' is required
 to back up recovery information),
 the existing command example on this page
 `manage-bde -protectors -adbackup C:.` is incorrect
 and needs to be replaced with a working backup command.

Thanks to rafalfitt for raising this issue and reporting the error.

Changes proposed:
- replace existing command sample with an extended PowerShell script
- change focus from drive C: to $env:SystemDrive in case C: is incorrect

Additional notes and caveats:
I do not have access to a test scenario to verify this content myself,
so I rely heavily on user feedback and confirmation from MS Docs team
members with regards to the suggested changes.

issue ticket closure or reference:

Closes #5159