Update credential-guard-protection-limits.md
Related to PR: https://github.com/MicrosoftDocs/windows-itpro-docs/pull/3789

Sorry for the delay, I have corrected the article to remove the video at the top and then at the bottom replace the video with the LinkedIn video relevant to this article.  I also put a note and link to the LinkedIn Learning subscription that is needed.