WHfB/Error mitigation: convert table to MarkDown
Description:

Based on community feedback in issue ticket #5455, the Error Code
0x801C044D is listed twice in the mitigation section, not adjacent, and
with different explanations. The suggestion is to group them together
to make it more obvious that there are 2 separate causes & mitigations.

Thanks to J.R. Jenkins (jrjenk) for reporting this issue.

In order to simplify editing and information management in the table, I
propose to change the table formatting from simple HTML to MarkDown.
I also suggest sorting the lines based on the Error code Hex values.

Changes proposed:
- convert the Error mitigations table to MarkDown
- rearrange the row lines and sort by Hex value (1 exception)
- place the 2 errors with the same Hex value adjacent to each other
- add some missing punctuation
- adjust spacing and remove some redundant whitespace

issue ticket closure or reference:

Closes #5455