Configure DC Certificates: Subject Name tab text
Description:
As reported and shown in issue ticket #5600 (Update text). one
of the tabs in Properties of New Template is referred to as Subject,
whereas the actual tab name is "Subject Name" in the doc page.

Thanks to Rami (drunkrhin0) for reporting this discrepancy.

Proposed change:

- add the word  Name to the "On the **Subject** tab" text (2x)

issue ticket closure or reference:

Closes #5600