Update white-glove.md
From testing, this seems to be the tech flow behavior regarding software deployment - additionally, the behavior we've seen is that the green screen appears once enrollment is done, not once everything is applied - this means the screen can go green but the reseal button won't appear until configuration and install is complete. 

Without ESP, the device enrolls, the screen goes green, the reseal button appears, but launching a command prompt and exploring the disk validates software didn't install. With ESP, the screen goes green within the same period, but there's a 20 minute delay after that before the reseal button appears - in this scenario, at reseal, the software is installed as expected.

Mostly just trying to get this behavior documented if it's correct, but if it's incorrect, some clarification on the software install behavior during tech flow would be good, because the expectation that software will install during the technician flow seems to not be consistent and this is the only reason I could find from the documentation that explains the behavior we're seeing.