Autopilot reset removes the device's primary user
https://docs.microsoft.com/en-us/intune/fundamentals/whats-new#week-of-july-15-2019
  
Windows Autopilot reset removes the device's primary user
When Autopilot reset is used on a device, the device's primary user will be removed. The next user who signs in after the reset will be set as the primary user. This feature will be rolled out to all customers over the next few days.