Deploy WDAC/Intune: update intro description
As discussed in issue ticket #4482 (Custom WDAC policy in Intune), it
would be useful to add a detail on implementing a custom WDAC policy
in Intune. Without this detail, the implication is that you can ONLY
use the Endpoint Protection template to configure WDAC in Intune.

Thank you to Air-Git for following up on this topic and the content.

Proposed change:

- add details missed in the previous PR #5659 (Content Update)

issue ticket closure or reference:

Ref. #4482 (already closed)