System Guard/Requirements: adjust description
Description:

As suggested in issue ticket #5515 (confusing phrasing about
requirements), the first sentence under the heading
"Requirements Met by System Guard Enabled Machines" can be interpreted
to mean that the act of enabling System Guard magically causes the
correct hardware to exist in the computer. This is an attempt to
adjust the sentence to be less ambiguous.

Thanks to marypcbuk for pointing out this ambiguity.

Proposed change:

- Adjust the sentence to state that the hardware requirements are
  already met if System Guard is enabled.

issue ticket closure or reference:

Closes #5515