Windows/Deployment: remove BGCOLOR from tables
Description:

As pointed out in the issue ticket #5819
(**bad color choice for table headers - dark & high contrast themes**),
bgcolor='#a0e4fa' is not a good choice for a dark theme page where the
black & white colors are swapped between text and background.

A similar issue is reported in ticket #5810.

The sensible choice here is to remove hardoded coloring and let this be
managed in the site CSS or other over-arching site scripting.

Proposed change:
- remove the color coding ` bgcolor='#a0e4fa'` from the tables

issue ticket closure or reference:

Closes #5819