Windows/Deployment: VAMT System Requirements
Description:

As per user feedback in issue ticket #5818 (Operating System
Requirements), it is desirable to allow for later versions of Windows
Server to be shown in the list of Operating Systems under "Minimum
system requirement".

There is also a visible defect in last row of the table, breaking the
last one and a half sentence over to the first column of the next row.

Thanks to laolive382 for posting this request and making me notice the
(ever so slightly) broken table.

Changes proposed:
- change "2008 R2, or Windows Server 2012" to
  "2008 R2, Windows Server 2012, or later"
- remove line break in the last row to restore the row structure
- add MarkDown spacing in the table for readability/editing

issue ticket closure or reference:

Closes #5818