Removed links pointing to a personal blog
I think this reference should be removed as the content is pointing to a non-official document. It's a personal blog and we can not refer any information from any personal blog in MS documentation.  All of the links pointing to a personal blog should be removed. Link to **Configuring a proxy for your Surface Hub** , **Configuring a proxy for your Surface Hub** and **Surface Hub and the Skype for Business Trusted Domain List** .

Problem: https://github.com/MicrosoftDocs/windows-itpro-docs/issues/5847