Removed the link
As the link is not pointing to a proper source and there is no other information is available on Microsoft docs so I have removed this link. The link provided was pointing to the LinkedIn course intro video and doesn't fit with the context of the article. 

Problem: https://github.com/MicrosoftDocs/windows-itpro-docs/issues/5854