Update linux-install-with-puppet.md
@mjcaparas 

Edit complete! Again, we're using "install" and "deploy" interchangeably. In most cases, I think "install" is more accurate and seems to be what's used in the code, but "deploy" has a slightly different meaning and might be more accurate in some instances, so it would be good to get crisper with the usage. 

TOC titles don't need to map verbatim to H1 headings, but if you stick with the TOC version, use "Puppet-based installation" (or Puppet-based deployment if switching to that language throughout).

On line 172: do you mean "2 if the device is not yet onboarded"?

Thanks!
Kelly