Resetting all references back to "3D Viewer Beta"
Someone changed all the references in this article to "3D Viewer," but that isn't correct. The app this article references is 3D Viewer Beta, which was discontinued several years ago and only works on HoloLens (1st gen). I've updated the article to clarify both, as it was confusing partners attempting to load FBX files on HoloLens 2.