MD-ATP/Troubleshoot onboarding: typo corrections
Description:

As reported in issue ticket #6378
(HKLM\SOFTWARE\Policies\Microsoft\Windows Advanced Threat is the wrong
path.), the Registry path in one of the table cells is incomplete due
to a typo. There is also a grammar typo in the same description.

This PR aims to resolve these issues.

Thanks to maskovli for reporting this issue.

Changes proposed:
- Add the missing word to the Registry path
- Change "was ran" to "has been run" (grammar)
- Whitespace correction: change a tab to 3 spaces

Ticket closure or reference:

Closes #6378