Applies to: Windows 10, Windows Server 2019
As per issue ticket #6441 (Supported windows versions are not valid),
this article incorrectly lists both deprecated and outdated OS
versions, both for Server and Client computers. The article itself
states that this policy was introduced in Windows 10, version 1703.

Thanks to yogeshasalkar for reporting this issue.

Changes proposed:
- Change the "Applies to" section to only Windows 10 and Server 2019
- Remove redundant end-of-line whitespace from 3 lines.

Ticket closure or reference:

Closes #6441