MDATP/Requirements: add Linux Public Preview notes
Description:

As noted in issue ticket #6430 (Linux support is in preview), this page
does not have a description of Linux support for MD-ATP being Public
Preview only for Linux at this time. This PR aims to rectify that
information detail.

Thanks to memildin (Mel) for reporting this issue.

Changes proposed:
- Add info comment behind "Linux" in the bullet point list
- Add a sentence at the end of the following Note blob
- Remove redundant phrase "that are" from Note to improve readability
- Add a missing plural S at the end of "Related topic" (2 links)
- Whitespace changes (might as well correct those)
    - add MarkDown indent marker compatibility spacing
    - remove redundant spacing in bullet point lists (only 1 needed)
    - remove redundant blank lines, 2 blank lines is sufficient spacing
    - remove redundant blanks at end-of-line (EOL whitespace)
    - replace a tab character with 4 blank spaces (GitHub codestyle)

Ticket closure or reference:

Closes #6430