BitLocker/MDT: TPM typo & link updates
Description:

As reported in issue ticket #6538 (TPM, not TMP), there is a typo where
the initialism TPM is misspelled as "TMP". Although a common variable
in the Microsoft Windows environment, it is not correct in this note.

Thanks to ntw2 for reporting this typo.

Further changes suggested: Replace 2 out of 3 fwlinks with their
current target page links (the 3rd is a rabbit hole too deep for now).

Changes proposed:
- "TMP" corrected to TPM (as it should be)
- fwlink/p/?LinkId=619548 -> docs.microsoft.com/previous-versions/
  windows/it-pro/windows-7/dd875529(v=ws.10) ((could use a new page))
- fwlink/p/?LinkId=167133 -> gallery.technet.microsoft.com/
  ScriptCenter/b4dee016-053e-4aa3-a278-3cebf70d1191
  ((marked for retirement, needs a backup or replacement))
- Remove redundant end-of-line whitespace for 2 lines

Additional notes:

Please feel free to suggest improved link replacements, especially for
the untouched fwlink, [Check to see if the TPM is enabled.]
(https://go.microsoft.com/fwlink/p/?LinkId=619549) This link lands on
the top blog archive menu: https://docs.microsoft.com/archive/blogs/ .

Ticket closure or reference:

Closes #6538