Update mac-install-with-intune.md
A note should be added to the very beginning of this article that stands out. This document explains the legacy method for deploying and configurating the Microsoft Defender ATP application on macOS devices. The note should explain this, and point admins to the new docs and the blog post that show the native experience which is much easier.  

The docs to link to from the note are -

Blog: https://techcommunity.microsoft.com/t5/microsoft-endpoint-manager-blog/microsoft-endpoint-manager-simplifies-deployment-of-microsoft/ba-p/1322995
Deployment: https://docs.microsoft.com/en-us/mem/intune/apps/apps-advanced-threat-protection-macos
Configuration: https://docs.microsoft.com/en-us/mem/intune/protect/antivirus-microsoft-defender-settings-macos

In order of steps in the note, the blog should come first, then config, then deployment. Thank you!