Identity Protection/VPN: grammar, links & spacing
As reported in issue ticket #6556 (Traffic filter sentence incomplete),
there is a missing part in the sentence "Network admins to effectively
add interface specific firewall rules on the VPN Interface." to make it
work as a full descriptive sentence in this context. This PR aims to
correct this issue, in addition to various other adjustments.

Thanks to klishb for reporting this issue.

Changes proposed:
- Add the missing part of the Traffic Filters sentence
- Update 2 outdated and permanently redirected MSDN links
- Uppercase adjustments for "Traffic filters" & "Lockdown"
- Add MarkDown indent marker compatibility spacing in the Note blob
- Reduce bullet point spacing from 3 to 1 in the "Applies to" section
- Remove all redundant end-of-line spacing
- Add missing space after the corrected sentence (after the period)

Ticket closure or reference:

Closes #6556