WDDG & WDCG HW readiness tool: typo & Arch Locale
Description:

As pointed out in ticket #5972 (closed) and #6574 (my ticket),
the script on this page stops working in localized versions of Windows
(Dutch mentioned in ticket #5972 specifically). This is caused by the
difference in how the strings "64-bit" and "32-bit" occur in non-English
localized versions of Windows.

Thanks to RvdHout (Ruud van den Hout) for suggesting the regex solution.

Changes proposed:
- $OSArch.Contains("64-bit") -> $OSArch -match ("64\-?\s?bits?")
- $OSArch.Contains("32-bit") -> $OSArch -match ("32\-?\s?bits?")
- Readiness Tool Version updated to 3.7.2 (thanks to RvdHout)
- Typo "archictecture" corrected to architecture (2 occurrences)
- Remove all redundant end-of-line (EOL) whitespace (blanks)
- Reduce number of blank lines before the signature block, from 3 to 2
- Add back a blank line between the metadata section and the page title

Ticket closure or reference:

Closes #6574

Ref. #5690 #5723 #6054 #6055
Ref. #5972 (closed)