Client Management/MDM: rename page names & titles
Description:

As requested by yannisle (Yannis Lempidakis) in issue ticket #6610 (Can
we change the title of this page to Poilcy "CSPs" supported by HoloLens
2? It makes it clear because the word Policy alone confuses customers),
this PR suggests to rename the page title and the file name according
to the request, to say "Policy CSPs supported by" instead of only
"Policies supported by", because that seems to confuse customers.

Thanks to yannisle for posting the request in ticket #6610.

Changes proposed:
- Rename the page titles, from "Policies supported by"[...]
  to "Policy CSPs supported by"[...], in 3 pages (so far)
- Rename the page names, from "policies-supported-by"[...]
  to "policy-csps-supported-by"[...], affecting 3 files (so far)
- Whitespace changes:
    - Reduce the bullet point link list spacing, from 3 to 1
    - Remove 1 redundant end-of-line space in the "div class" section
    - Add NewLine (line break) at the end-of-file (EOF) for all 3 files

Additional notes:
- [ ] Confirm whether 6 more files need the same change:
    - Policies supported by Group Policy
    - ADMX-backed policies
    - Policies supported by Windows 10 IoT Enterprise
    - Policies supported by Windows 10 IoT Core
    - Policies supported by Microsoft Surface Hub
    - Policies that can be set using Exchange Active Sync (EAS)

(See #6610#issuecomment-623102343 for details)

Ticket closure or reference:

Closes #6610