VPN/SSO NDES: SCEP link URL update
Description:

As reported in issue ticket #6766 (Dead link), the current link to
"Configure certificate infrastructure for SCEP" returns a 404 error
because the page has been moved to a new directory structure in commit
4ae71ed25d10 on February 28, 2020. 1702 files and folders have been
moved and/or renamed in this process, making it a game of chance to
correct more than one 404 link discovery at a time. (I hope some links
were corrected back then, but I have not researched the details.)

Thanks to rossmpersonal for reporting this 404 issue.

Changes proposed:

- Update the link URL to its current location
  (docs.microsoft.com/mem/intune/protect/certificates-scep-configure)
- Remove redundant end-of-line whitespace (blanks) throughout the page

Ticket closure or reference:

Closes #6766