User credential preferred
Even though Device Credential is an option on the GPO, the device credential gives error while auto-enrollment tasks running through the Task Scheduler. To avoid this error we need to choose the User Credential option from the dropdown to auto-enroll the device. The below line has been updated on the document.
5. Click **Enable**, and select **User Credential** from the dropdown **Select Credential Type to Use**, then click **OK**.