Edit pass for Surface Book 3 Overview DO NOT MERGE
Light copy edits for surface-book-quadro.md.

Notes: Changed sentences in Table 3.
For Adobe Substance Designer, edited confusing sentence by removing "powered" and replacing it with "that is". The author should verify that the new sentence is acceptable.

For Autodesk 3ds Max, removed asterisks from second bullet. Not sure if this was a typo or meant to imply that the following text was a note.