Set up AADJ PC RDC: Note reference "UPN attribute"
Description:

As noted in issue ticket "FirstnameLastname reference doesn't resolve"
(#6902), there is a disparity between the cmdlet parameter detail
"the-UPN-attribute-of-your-user" and the explanatory use of the phrase
"FirstnameLastname" which is not used anywhere else on this page.

Thanks to tomten for reporting this issue.

Changes proposed:
- Change "FirstnameLastname" to "the-UPN-attribute-of-your-user"
- Add MarkDown codeblock fencing with PowerShell syntax name
- Reduce excessive usage of "then", "that" or "that are/that have"
- Replace "that have" with "using" where grammatically appropriate
- Extend "using:" to "by using any of the following:" (grammar fluency)
- Whitespace and codestyle adjustments:
    - Add MarkDown indent marker compatibility spacing where needed
    - Add blank line before bullet point lists, where needed
    - Remove redundant end-of-line blanks (4 lines)
    - Normalize MD indent marker indentation from 5 to 4 spaces
    - Normalize bullet point spacing from 3 to 1 (1 line)
    - Remove 10 redundant blank lines at the end of the page

Additional note: I am looking forward to feedback and suggestions for
my grammar changes as well as the 'net localgroup' cmdlet phrasing.

Ticket closure or reference:

Closes #6902