Updated guidance on using Windows in FIPS mode
Updated the "Using Windows in a FIPS 140-2 approved mode of operation" to be up to date with current Microsoft recommendations and added in additional question to the FAQ to give clarity on the compliance status of SMB3 in Windows. Also re-added the fips@microsoft contact information because it was dropped off in a previous update to this page.