Update kernel-dma-protection-for-thunderbolt.md
Add the following FAQ:
### My system's Kernel DMA Protection is off. Can DMA-remapping for a specific device be turned on?

Yes. DMA remapping for a specific device can be turned on independent from Kernel DMA Protection. (i.e. if the driver opts in and VT-d is turned on, then DMA remapping will be enabled for the devices driver, even if Kernel DMA Protection is off).

Kernel DMA Protection is a policy that allows/disallows devices to perform DMA, based on their remapping state/capabilities.