Update windows-10-subscription-activation.md
We get a fair number of cases in support regarding Windows 10 Enterprise Subscription Based Activation and GCC/GCC High tenants. At this time it is not supported for GCC and GCC High Azure tenants. I would like this doc edited to reflect that so that we have a place to direct customers when this request comes up.