Added a Store Content Related URL
As the user has mentioned that to access the content of the store required URL was missing. I have updated the URL.

Problem: https://github.com/MicrosoftDocs/windows-itpro-docs/issues/8874

Reset of the URL's mentioned in the comment doesn't have any reference in MS docs and previous version document. So as of now, I have dropped them. 

Thanks.
Imran.