Make verifier run on next boot only
In its current state, the verifier does not spcify a /bootmode, which causes it to use the default behavior, "persistent." As a result, on an incompatible hardware configuration, the system may get stuck in a boot loop. (Ask me how I know!)

This change specifies that the verifier should only run on the next boot, preventing a user from getting their computer stuck in a situation that Windows cannot resolve on its own.