More context on config script
Config script docs inaccurately said that the script configures GP directly. It configures regkeys. This means that if GP Policy X is set to 1, but the regkey value is supposed to be 0, then the config script will set the regkey to 0; however, on the next GP refresh, the regkey will be set back to 1.